로컬 환경 임시저장2

+++ src/main/java/com/sparta/itsmine/domain/kakaopay/service/KakaoPayService.java
@@ -306,7 +306,7 @@ public class KakaoPayService {
                 KakaoPayTid kakaoPayTid = kakaoPayRepository.findByAuctionId(auction.getId());
                 kakaoCancel(kakaoPayTid.getTid());
                 auctionRepository.delete(auction);
-            } else{ //if(auction.getStatus().equals(NEED_PAY)) {
+            } else if(auction.getStatus().equals(NEED_PAY)) {
                 auctionRepository.delete(auction);
             }
         }
